requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
Brotli==1.0.9
click==8.1.3
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
whitenoise==6.4.0
